backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM & AI
openai>=1.12.0
ollama>=0.1.6
langchain>=0.1.6
langchain-community>=0.0.19

# Speech Processing (Whisper)
openai-whisper>=20231117
torch>=2.2.0
torchaudio>=2.2.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# NLP & Text Analysis
spacy>=3.7.0
textstat>=0.7.3
language-tool-python>=2.7.1

# Gramformer - Install separately via:
# pip install git+https://github.com/PrithivirajDamodaran/Gramformer.git

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
supabase>=2.3.0

# AWS
boto3>=1.34.0
aws-lambda-powertools>=2.31.0

# Utilities
python-dotenv>=1.0.1
httpx>=0.26.0
aiofiles>=23.2.1
numpy>=1.26.0
pandas>=2.2.0

# WebSocket support (included with uvicorn[standard])
websockets>=12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Image processing
Pillow>=10.0.0
huggingface_hub>=0.20.0